requirements.txt
streamlit==1.34.0
pdfminer.six==20231228
langchain==0.1.14
transformers==4.40.2
sentence-transformers==2.7.0
faiss-cpu==1.8.0
tiktoken==0.6.0
contractions==0.1.73
nltk==3.8.1
translate==3.6.1
deep-translator==1.11.4
sentencepiece==0.2.0